implement context module functions pattern
