+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled, { ThemeProvider } from 'styled-components'
 import { GlobalStyles } from './components/GlobalStyles'
 import { theme } from './lib/theme'
-import { EnrollmentProvider } from './context/enrollment'
+import { EnrollmentProvider } from './context/enrollment-context'
 import FormSteps from './components/FormSteps'
 import Typography from './components/Typography'
 import { SITE_TITLE } from './lib/constants'
